tests/dialer/test_contact_intent.html

<!DOCTYPE HTML>
<html>
  <head>
    <title>Dialer contact intent</title>
    <script type="text/javascript" src="/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="/SimpleTest/test.css" />
  </head>
  <body>
    <p id="display"></p>
    <div id="content" style="display: block">
      <iframe src="/tests/gaia/apps/dialer/dialer.html?choice=contact"></iframe>
    </div>
    <pre id="test">
    <script class="testbody" type="text/javascript">
      var dialerFrame = document.getElementsByTagName('iframe')[0];

      dialerFrame.addEventListener('load', function runWhenLoaded(evt) {
        dialerFrame.removeEventListener('load', runWhenLoaded);

        var dialerWindow = dialerFrame.contentWindow;
        // TODO: trigger a real visibility change event when the webapi fixe lands
        dialerWindow.visibilityChanged(dialerWindow.location.href);

        ok(!dialerWindow.document.getElementById('contacts-view').hidden, 'Contact view displayed');
        ok(dialerWindow.Contacts._loaded, 'Contacts loaded');
      });
    </script>
    </pre>
  </body>
</html>
